feat: autocommand preview does not work
Unable to respond to an interaction while returning auto complete
result. The idea of having a preview, an empheral message that displays
the emote while the user types is not possible for now.

Empheral messages can only be sent when the user triggers a slash
command, due to security reasons...

https://stackoverflow.com/questions/68298805/send-spontaneous-ephemeral-discord-message