fix(commands): correct handling of `Command.InteractionRespond` and `Command.EditInteractionResponse`
`InteractionRespond` allows the passed in function to respond with an
itneraction response. If the command fails, the error will be sent as an
ephemeral interaction response.

`EditInteractionResponse` sends a defered reply, and expects the passed
in function to edit the reply. If the command fails, the error will be
sent as an EditInteractionResponse. Due to the nature of editing a
response, the response will not be ephimeral. So the error message will
be deleted in 5 seconds.